requirements.txt
Django==3.1.3
django-cors-headers==3.7.0
django-heroku==0.3.1
django-knox-rest==4.1.0
djangorestframework==3.12.2
gunicorn==20.1.0
Pillow==8.1.0